+++ lib/ssr/deno.rb
@@ -2,9 +2,9 @@
 
 require 'json'
 require_relative 'deno/version'
-
-# Load the native Rust extension (compiled by rb-sys / rake-compiler)
 require_relative 'deno/ssr_deno'
+require_relative 'deno/manifest'
+
 require_relative 'deno/instrumenter'
 require_relative 'deno/bundle'